# backend/api/swagger/swagger.yaml
swagger: '2.0'
info:
  title: "Open Referral API"
  description: "This is a specification describing the Open Referral API, using the Human Services Specification. This has been extracted from the Ohana API deployment, but will be standardized to refer to the universal standard that will apply to any Open Referral compliant API."
  termsOfService: "https://openreferral.org/documentation/"
  contact:
    name: Open Referral
    url: https://openreferral.org/
    email: info@openreferral.org
  license:
    name: Open Data Commons Public Domain
    url: http://opendatacommons.org/licenses/pddl/
  version: v1.1
host: localhost:3000
basePath: /
schemes:
- http
produces:
- application/json
paths:
  /contacts/:
    x-swagger-router-controller: contacts
    get:
      summary: Get Contacts
      description: Get Contacts
      operationId: listContacts
      parameters:
        - in: query
          name: query
          required: false
          type: array
          description: A query to filter results by (e.g. <field>==<value>). Multiple entries are allowed.
          collectionFormat: multi
          items:
            type: string
        - in: query
          type: number
          name: page
          description: The particular page of results.
        - in: query
          type: number
          name: per_page
          description: Amount of locations to return per page, up to 100.
      responses:
        '200':
          description: Contact Response
          schema:
            type: array
            items:
              $ref: "#/definitions/contact"
      tags:
      - Contacts
    post:
      summary: Add Contact
      description: Add contact
      operationId: addContact
      parameters:
        - in: body
          name: body
          schema:
            $ref: '#/definitions/contact'
      security:
        - appid: []
        - appkey: []
      responses:
        '200':
          description: Contact Response
          schema:
            type: array
            items:
              $ref: "#/definitions/contact"
      tags:
      - Contacts
  /contacts/{contact_id}/:
    x-swagger-router-controller: contacts
    get:
      summary: Get Contact
      description: Get Contact
      operationId: getContact
      parameters:
        - in: path
          required: true
          type: string
          name: contact_id
          description: 'The contact id.'
      responses:
        '200':
          description: Contact Response
          schema:
            type: array
            items:
              $ref: "#/definitions/contact"
      tags:
      - Contacts
    put:
      summary: Update Contact
      description: Update an existing contact.
      operationId: updateContact
      parameters:
        - in: path
          required: true
          type: string
          name: contact_id
          description: 'The contact id.'
        - in: body
          name: body
          schema:
            $ref: '#/definitions/contact'
      security:
        - appid: []
        - appkey: []
      responses:
        '200':
          description: Contact Response
          schema:
            type: array
            items:
              $ref: "#/definitions/contact"
      tags:
      - Contacts
    delete:
      summary: Delete Contact
      description: Delete Contact
      operationId: deleteContact
      parameters:
        - in: path
          required: true
          type: string
          name: contact_id
          description: 'The contact id.'
      tags:
      - Contacts
      security:
      - appid: []
      - appkey: []
      responses:
        '200':
          description: Contact Response
          schema:
            type: array
            items:
              $ref: "#/definitions/contact"
  /contacts/{contact_id}/phones/:
    x-swagger-router-controller: contacts
    get:
      summary: Get Phones
      description: Get Phones
      operationId: listContactPhones
      parameters:
        - in: path
          required: true
          type: string
          name: contact_id
          description: 'The unique contact id.'
      responses:
        '200':
          description: Phone Response
          schema:
            type: array
            items:
              $ref: "#/definitions/phone"
      tags:
      - Contacts
    post:
      summary: Add Phone
      description: Add Phone
      operationId: addContactPhone
      parameters:
        - in: path
          required: true
          type: string
          name: contact_id
          description: 'The unique contact id.'
        - in: body
          name: body
          schema:
            $ref: '#/definitions/phone'
      security:
        - appid: []
        - appkey: []
      responses:
        '200':
          description: Phone Response
          schema:
            type: array
            items:
              $ref: "#/definitions/phone"
      tags:
      - Contacts
  /contacts/{contact_id}/phones/{phone_id}/:
    x-swagger-router-controller: contacts
    get:
      summary: Get Phone
      description: get Phone
      operationId: getContactPhone
      parameters:
        - in: path
          required: true
          type: string
          name: contact_id
          description: The unique contact id.
        - in: path
          required: true
          type: string
          name: phone_id
          description: The unique phone id.
      responses:
        '200':
          description: Phone Response
          schema:
            type: array
            items:
              $ref: "#/definitions/phone"
      tags:
      - Contacts
    put:
      summary: Update Phone
      description: Update Phone
      operationId: updateContactPhone
      parameters:
        - in: path
          required: true
          type: string
          name: contact_id
          description: 'The unique contact id.'
        - in: path
          required: true
          type: string
          name: phone_id
          description: 'The unique phone id.'
        - in: body
          name: body
          schema:
            $ref: '#/definitions/phone'
      security:
        - appid: []
        - appkey: []
      responses:
        '200':
          description: Phone Response
          schema:
            type: array
            items:
              $ref: "#/definitions/phone"
      tags:
      - Contacts
    delete:
      summary: Delete Phone
      description: Delete Phone
      operationId: deleteContactPhone
      parameters:
        - in: path
          required: true
          type: string
          name: contact_id
          description: 'The unique contact id.'
        - in: path
          required: true
          type: string
          name: phone_id
          description: 'The unique phone id.'
      security:
        - appid: []
        - appkey: []
      responses:
        '200':
          description: Phone Response
          schema:
            type: array
            items:
              $ref: "#/definitions/phone"
      tags:
      - Contacts
...
securityDefinitions:
  appid:
    type: apiKey
    in: header
    name: x-appid
  appkey:
    type: apiKey
    in: header
    name: x-appkey
definitions:
...
  contact:
    properties:
      id:
        description: 'Each contact must have a unique identifier.'
        type: string
      organization_id:
        description: 'The identifier of the organization for which this is a contact.'
        type: string
      service_id:
        description: 'The identifier of the service for which this is a contact.'
        type: string
      service_at_location_id:
        description: 'The identifier of the ‘service at location’ table entry, when this contact is specific to a service in a particular location.'
        type: string
      name:
        description: 'The name of the person.'
        type: string
      title:
        description: 'The job title of the person.'
        type: string
      department:
        description: 'The department that the person is part of.'
        type: string
      email:
        description: 'The email address of the person.'
        type: string
    required:
      - id
...
  phone:
    properties:
      id:
        description: 'Each entry must have a unique identifier.'
        type: string
      location_id:
        description: 'The identifier of the location where this phone number is located.'
        type: string
      service_id:
        description: 'The identifier of the service for which this is the phone number.'
        type: string
      organization_id:
        description: 'The identifier of the organisation for which this is the phone number.'
        type: string
      contact_id:
        description: 'The identifier of the contact for which this is the phone number.'
        type: string
      service_at_location_id:
        description: 'The identifier of the ‘service at location’ table entry, when this phone number is specific to a service in a particular location.'
        type: string
      number:
        description: 'The phone number.'
        type: string
      extension:
        description: 'The extension of the phone number.'
        type: ['number', 'null']
      type:
        description: 'Whether the phone number relates to a fixed or cellular phone.'
        type: string
      department:
        description: 'The department for which this is the phone number.'
        type: string
      language:
        description: 'A comma separated list of ISO 639-1, or ISO 639-2 [language codes](available at http://www.loc.gov/standards/iso639-2/php/code_list.php) to represent the languages available from this phone service. The three-letter codes from ISO 639-2 provide greater accuracy when describing variants of languages, which may be relevant to particular communities.'
        type: string
      description:
        description: 'A description providing extra information about the phone service (e.g. any special arrangements for accessing, or details of availability at particular times.'
        type: string
    required:
      - id
      - number
...